+++ src/apps/gearbox/contracts/index.ts
@@ -5,7 +5,9 @@ import { ContractFactory } from '~contract/contracts';
 import { Network } from '~types/network.interface';
 
 import { ContractsRegister__factory } from './ethers';
+import { CreditManagerV2__factory } from './ethers';
 import { DieselToken__factory } from './ethers';
+import { PhantomToken__factory } from './ethers';
 import { PoolService__factory } from './ethers';
 
 // eslint-disable-next-line
@@ -20,8 +22,14 @@ export class GearboxContractFactory extends ContractFactory {
   contractsRegister({ address, network }: ContractOpts) {
     return ContractsRegister__factory.connect(address, this.appToolkit.getNetworkProvider(network));
   }
+  creditManagerV2({ address, network }: ContractOpts) {
+    return CreditManagerV2__factory.connect(address, this.appToolkit.getNetworkProvider(network));
+  }
   dieselToken({ address, network }: ContractOpts) {
     return DieselToken__factory.connect(address, this.appToolkit.getNetworkProvider(network));
+  }
+  phantomToken({ address, network }: ContractOpts) {
+    return PhantomToken__factory.connect(address, this.appToolkit.getNetworkProvider(network));
   }
   poolService({ address, network }: ContractOpts) {
     return PoolService__factory.connect(address, this.appToolkit.getNetworkProvider(network));
@@ -29,5 +37,7 @@ export class GearboxContractFactory extends ContractFactory {
 }
 
 export type { ContractsRegister } from './ethers';
+export type { CreditManagerV2 } from './ethers';
 export type { DieselToken } from './ethers';
+export type { PhantomToken } from './ethers';
 export type { PoolService } from './ethers';