+++ src/apps/gearbox/gearbox.definition.ts
@@ -8,15 +8,30 @@ export const GEARBOX_DEFINITION = appDefinition({
   name: 'Gearbox',
   description: 'Composable Leverage Protocol',
   url: 'https://gearbox.fi ',
+
   groups: {
     lending: {
       id: 'lending',
       type: GroupType.TOKEN,
       label: 'Lending',
     },
+
+    creditAccounts: {
+      id: 'credit-accounts',
+      type: GroupType.POSITION,
+      label: 'Credit Accounts',
+    },
+
+    phantom: {
+      id: 'phantom',
+      type: GroupType.TOKEN,
+      label: 'Phantom Tokens',
+    },
   },
+
   tags: [AppTag.COLLATERALIZED_DEBT_POSITION, AppTag.FARMING, AppTag.LENDING],
   keywords: [],
+
   links: {
     discord: 'https://discord.com/invite/gearbox',
     github: 'https://github.com/Gearbox-protocol',