+++ src/apps/mean-finance/mean-finance.definition.ts
@@ -30,6 +30,7 @@ export const MEAN_FINANCE_DEFINITION = appDefinition({
     [Network.POLYGON_MAINNET]: [AppAction.VIEW],
     [Network.OPTIMISM_MAINNET]: [AppAction.VIEW],
     [Network.ARBITRUM_MAINNET]: [AppAction.VIEW],
+    [Network.ETHEREUM_MAINNET]: [AppAction.VIEW],
   },
 
   primaryColor: '#3076F6',