+++ src/apps/pendle-v2/contracts/index.ts
@@ -7,6 +7,7 @@ import { Network } from '~types/network.interface';
 import { PendleMarket__factory } from './ethers';
 import { PendleMarketFactory__factory } from './ethers';
 import { PendlePrincipalToken__factory } from './ethers';
+import { PendleVotingEscrow__factory } from './ethers';
 import { PendleYieldContractFactory__factory } from './ethers';
 import { PendleYieldToken__factory } from './ethers';
 import { StandardizedYield__factory } from './ethers';
@@ -29,6 +30,9 @@ export class PendleV2ContractFactory extends ContractFactory {
   pendlePrincipalToken({ address, network }: ContractOpts) {
     return PendlePrincipalToken__factory.connect(address, this.appToolkit.getNetworkProvider(network));
   }
+  pendleVotingEscrow({ address, network }: ContractOpts) {
+    return PendleVotingEscrow__factory.connect(address, this.appToolkit.getNetworkProvider(network));
+  }
   pendleYieldContractFactory({ address, network }: ContractOpts) {
     return PendleYieldContractFactory__factory.connect(address, this.appToolkit.getNetworkProvider(network));
   }
@@ -43,6 +47,7 @@ export class PendleV2ContractFactory extends ContractFactory {
 export type { PendleMarket } from './ethers';
 export type { PendleMarketFactory } from './ethers';
 export type { PendlePrincipalToken } from './ethers';
+export type { PendleVotingEscrow } from './ethers';
 export type { PendleYieldContractFactory } from './ethers';
 export type { PendleYieldToken } from './ethers';
 export type { StandardizedYield } from './ethers';