+++ src/apps/penguin/penguin.module.ts
@@ -1,7 +1,6 @@
 import { Register } from '~app-toolkit/decorators';
 import { AbstractApp } from '~app/app.dynamic-module';
 
-import { AvalanchePenguinBalanceFetcher } from './avalanche/penguin.balance-fetcher';
 import { AvalanchePenguinChefV1FarmContractPositionFetcher } from './avalanche/penguin.chef-v1-farm.contract-position-fetcher';
 import { AvalanchePenguinChefV2FarmContractPositionFetcher } from './avalanche/penguin.chef-v2-farm.contract-position-fetcher';
 import { AvalanchePenguinIPefiTokenFetcher } from './avalanche/penguin.i-pefi.token-fetcher';
@@ -16,7 +15,6 @@ import { PenguinAppDefinition, PENGUIN_DEFINITION } from './penguin.definition';
   providers: [
     PenguinAppDefinition,
     PenguinContractFactory,
-    AvalanchePenguinBalanceFetcher,
     AvalanchePenguinChefV1FarmContractPositionFetcher,
     AvalanchePenguinChefV2FarmContractPositionFetcher,
     AvalanchePenguinIPefiTokenFetcher,